refactor:t-auth: Delete RoleName, RoleDto; fix:t-auth: Prohibit the removal of ADMIN, USER roles;
