+++ t-auth/src/main/java/com/pavbatol/talentium/jwt/JwtAuthentication.java
@@ -1,6 +1,6 @@
 package com.pavbatol.talentium.jwt;
 
-import com.pavbatol.talentium.role.model.RoleName;
+import com.pavbatol.talentium.shared.auth.model.RoleName;
 import io.jsonwebtoken.Claims;
 import lombok.*;
 import org.springframework.security.core.Authentication;
