+++ t-auth/src/main/java/com/pavbatol/talentium/user/dto/UserDtoUpdate.java
@@ -1,6 +1,6 @@
 package com.pavbatol.talentium.user.dto;
 
-import com.pavbatol.talentium.role.dto.RoleDto;
+import com.pavbatol.talentium.shared.auth.dto.RoleDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
