+++ t-auth/src/main/java/com/pavbatol/talentium/user/service/UserServiceImpl.java
@@ -6,11 +6,11 @@ import com.pavbatol.talentium.app.util.Checker;
 import com.pavbatol.talentium.email.service.EmailService;
 import com.pavbatol.talentium.jwt.JwtProvider;
 import com.pavbatol.talentium.role.model.Role;
-import com.pavbatol.talentium.role.model.RoleName;
 import com.pavbatol.talentium.role.storage.RoleRepository;
 import com.pavbatol.talentium.shared.auth.dto.RoleDto;
 import com.pavbatol.talentium.shared.auth.dto.UserDtoUpdateInsensitiveData;
 import com.pavbatol.talentium.shared.auth.dto.UserDtoUpdateSensitiveData;
+import com.pavbatol.talentium.shared.auth.model.RoleName;
 import com.pavbatol.talentium.user.Verification.model.VerificationToken;
 import com.pavbatol.talentium.user.Verification.service.VerificationTokenService;
 import com.pavbatol.talentium.user.Verification.storage.VerificationTokenRepository;
